% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ms.r
\name{call_ms}
\alias{call_ms}
\title{Call MS from R}
\usage{
call_ms(nsam, howmany, cmd, verbose = TRUE)
}
\arguments{
\item{nsam}{number of samples (gametes) to draw}

\item{howmany}{how many replicates to run}

\item{cmd}{the command to pass to MS}
}
\description{
Call Hudon's MS from R, returning a string of results (that can be parsed
with parse_ms()).
}

