% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camera}
\alias{camera}
\title{camera}
\format{
A data frame with 612 rows and 5 variables:
\describe{
\item{id}{Indicates ID of respondent}
\item{memory_format}{Indicates type of memory format: "Memory Stick", "SD Card", or "Compact Flash Card"}
\item{lens}{Indicates type of camera lens: "Glass Lens 8x Zoom", "Plastic Lens 8x Zoom", or "Plastic Lens 4x Zoom"}
\item{price}{Indicates price of camera: "300 euro", "100 euro", "200 euro"}
\item{ratings}{Indicates the rating of the camera, value from 1-100}
}
}
\source{
{Compiled by Xi Chen}
}
\usage{
camera
}
\description{
Contains camera data used in this course.
}
\examples{
data(camera)
}
\keyword{datasets}
