% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_mean.R
\name{cluster_mean}
\alias{cluster_mean}
\title{cluster mean}
\usage{
cluster_mean(anova_cluster)
}
\arguments{
\item{anova_cluster}{an R ANOVA analysis object that uses the clustering outcome from a cluster analysis as the factors and the attributes that are used for the clustering as the dependent variables.}
}
\value{
mean_table a table of cluster means, each value is the mean of an attribute for a cluster.
}
\description{
This function gives you the centers (means) of all the attributes that are used for clustering for each cluster. You will get a table of cluster means.
}
\examples{
NA
}
