% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dim_toothpaste}
\alias{dim_toothpaste}
\title{dim toothpaste}
\format{
A matrix with 10 rows and 3 variables:
\describe{
\item{V1}{The first dimension or underlying attribute of the toothpaste brands}
\item{V2}{The second dimension or underlying attribute of the toothpaste brands}
\item{V3}{The third dimension or underlying attribute of the toothpaste brands}
}
}
\source{
{Compiled by Xi Chen}
}
\usage{
dim_toothpaste
}
\description{
Contains a matrix with the number of dimensions for toothpaste data used in this course.
}
\examples{
data(dim_toothpaste)
}
\keyword{datasets}
