% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_toothpaste}
\alias{dist_toothpaste}
\title{dist toothpaste}
\format{
A data frame with 10 row and 10 variables, with each row and column corresponding to a brand:
\describe{
\item{AquaFresh}{Distances of other brands to AquaFresh}
\item{Crest}{Distances of other brands to Crest}
\item{Colgate}{Distances of other brands to Colgate}
\item{Aim}{Distances of other brands to Aim}
\item{Gleem}{Distances of other brands to Gleem}
\item{PlusWhite}{Distances of other brands to PlusWhite}
\item{UltraBrite}{Distances of other brands to UltraBite}
\item{CloseUp}{Distances of other brands to CloseUp}
\item{Pepsodent}{Distances of other brands to Pepsodent}
\item{Sensodyne}{Distances of other brands to Sensodyne}
}
}
\source{
{Compiled by Xi Chen}
}
\usage{
dist_toothpaste
}
\description{
Contains a data frame that is a distance matrix between the 10 toothpaste brands discussed in this course.
}
\examples{
data(dist_toothpaste)
}
\keyword{datasets}
