% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_plot.R
\name{elbow_plot}
\alias{elbow_plot}
\title{elbow plot}
\usage{
elbow_plot(height, nclust = 10)
}
\arguments{
\item{height}{the "height" from the hclust object. height is a vector of within-cluster variation at different stage of the agglomeration procedure. It starts with the initial stage that each and every consumer is in its own cluster, and gradually combine users and clusters to the end stage that all consumers are in 1 cluster. Note that the values in height is always in an ascending order.}

\item{nclust}{the max number of clusters that you want to check. The default value is 10. You can change to other integer values if you want.}
}
\value{
An elbow plot which plots the within cluster variation (height) against the no. of clusters.
}
\description{
This function produces an elbow plot based the hierarchical clustering results (an R "hclust" object).
}
\examples{
NA
}
