% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bass.R
\name{estimate_bass}
\alias{estimate_bass}
\title{estimate_bass}
\usage{
estimate_bass(x)
}
\arguments{
\item{x}{the no. of adoptions at each time period in a chronological order. For example, x can be the no. of iPhones sold at each quarter of the year.}
}
\value{
The function returns a vector of bass parameters in the order of p, q, M, i.e., the innovation parameter p, the imitation parameter q and the market potential M.
}
\description{
This function estimates the bass parameters, given the no. of adoptions (n) at each time period.
}
\examples{
NA
}
