% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partworths.R
\name{get_partworths}
\alias{get_partworths}
\title{get partworths}
\usage{
get_partworths(mdl)
}
\arguments{
\item{mdl}{The R linear regression ( lm(.) ) object, i.e., the result of a dummy variable regression of ratings on the level dummies for a conjoint analysis}
}
\value{
a list that contains partworths values; each member of the list corresponds to an attribute; each member is a vector that contains the partworths of the levels of an attribute.
}
\description{
This function returns the partworths of levels from a rating-based conjoint analysis. The returned value is a list containing the partworths values of various levels.
}
\examples{
NA
}
