% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iphone}
\alias{iphone}
\title{iPhone data}
\format{
A data frame with 37 rows and 2 variables:
\describe{
\item{time}{Each quarter of the year}
\item{no_of_adoptions}{The number of adoptions (units sold)}
}
}
\source{
{Compiled by Xi Chen}
}
\usage{
iphone
}
\description{
Contains the units of sales of iPhone at each quarter of the year.
}
\examples{
data(iphone)
}
\keyword{datasets}
