% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marks_spencer}
\alias{marks_spencer}
\title{Marks Spencer}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{Week}{Indicates the no. of weeks, in total 100 weeks}
\item{Advertising}{The advertising intensity}
\item{Promotion}{a factor of two-levels - “Yes” or “No”; the baseline is already set to “No”}
\item{Price}{the price index}
\item{Brand_Equity}{a measure of weekly brand equity of Marks & Spencer}
\item{Sales}{the weekly sales}
}
}
\source{
{Compiled by Xi Chen}
}
\usage{
marks_spencer
}
\description{
Contains Marks Spencer data used in the Market Response Model sessions.
}
\examples{
data(marks_spencer)
}
\keyword{datasets}
