% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osn}
\alias{osn}
\title{osn}
\format{
A data frame with 120 rows and 6 variables of users from an online social network (OSN):
\describe{
\item{intrinsic}{The intrinsic preferences of users towards the OSN}
\item{habit}{To what extent users form a persistent habit of using the OSN}
\item{si_followers}{How susceptible users are to the influence of their followers}
\item{si_friends}{How susceptible users are to the influence of their friends}
\item{n_followers}{The (normalized) no. of followers of users}
\item{n_friends}{The (normalized) no. of friends of users}
}
}
\source{
{Compiled by Xi Chen}
}
\usage{
osn
}
\description{
Contains a data frame used for the Cluster Analysis session.
}
\examples{
data(osn)
}
\keyword{datasets}
