% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ms.r
\name{parse_keyvals}
\alias{parse_keyvals}
\title{Parse MS's key/value pairs, e.g. segsites and positions
returning the values as list if there are more than one}
\usage{
parse_keyvals(x)
}
\description{
Parse MS's key/value pairs, e.g. segsites and positions
returning the values as list if there are more than one
}
\keyword{internal}

