% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ms.r
\name{parse_ms}
\alias{parse_ms}
\title{Parse MS output from R}
\usage{
parse_ms(x)
}
\arguments{
\item{x}{character vector output from MS

Parse MS output from a character vector of lines of output to a tidy tibble.
Each replicate is a row, with a list-column for positions and a matrix of
gamete states.


You can use the purrr package to write map functions to summarize the
list-column of gametes.

importFrom(magrittr,"%>%")}
}
\description{
Parse MS output from R
}
\examples{
call_ms(10, 10, "-t 5") \%>\% parse_ms() 
}

