% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptual_map.R
\name{perceptual_map}
\alias{perceptual_map}
\title{Creating a perceptual map}
\usage{
perceptual_map(mds, dims)
}
\arguments{
\item{mds}{the result from the classic multi-dimensional scaling (cmdscale). It should be a matrix of dimensions N (no. of brands) by K (the no. of dimensions).}

\item{dims}{a vector that indicates which two dimensions that you want to plot. For example, dims = c(1,2) means you want to plot dimension 1 and 2, with Dim 1 as the X-axis and Dim 2 as the Y-axis.}
}
\value{
A perceptual map indicating the positions of brands in the two dimensions specified in dims.
}
\description{
This function creates a perceptual map by plotting any two dimensions of the multi-dimensional scaling.
}
\examples{
NA
}
