% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bass.R
\name{predict_bass}
\alias{predict_bass}
\title{predict_bass}
\usage{
predict_bass(T, bass.par)
}
\arguments{
\item{T}{The no. of time periods of your predictions.}

\item{bass.par}{The bass parameters in the order of ($p$,$q$,$M$) it outputs the cumulative no. of adoptions during $T$.}
}
\value{
The cumulative no. of adoptions ($N$) from time period 1 to time period $T$.
}
\description{
The function "predict_bass" is to predict the cumulative no. of adoptions ($N$) using the Bass parameters $p$, $q$ and $M$. From the prediction, you can get the adoption curve.
}
\examples{
NA
}
