% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{red_bull_ads}
\alias{red_bull_ads}
\title{Red Bull ads}
\format{
A data frame with 171 rows and 2 variables:
\describe{
\item{time}{Each day}
\item{no_of_adoptions}{The daily number of adoptions (views))}
}
}
\source{
{Compiled by Xi Chen}
}
\usage{
red_bull_ads
}
\description{
Contains the daily views of Red Bull ads.
}
\examples{
data(red_bull_ads)
}
\keyword{datasets}
