% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{sample_stats}
\alias{sample_stats}
\title{Mutate a tibble of results from parse_ms(), adding summary statistics}
\usage{
sample_stats(x, .n)
}
\arguments{
\item{x}{tibble of results from parse_ms()}

\item{.n}{number of samples}
}
\description{
Mutate a tibble of results from parse_ms(), adding summary statistics
}
