% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree_plot.R
\name{scree_plot}
\alias{scree_plot}
\title{scree plot}
\usage{
scree_plot(dist, ndim = 10)
}
\arguments{
\item{dist}{a distance matrix with each element as the distance between two brands. The own distance is set to zero. The distance from Brand 1 to Brand 2 is the same as that from Brand 2 to 1. "dist" is a symmetric matrix with diagonal elements equal to zeros.}

\item{ndim}{the no. of dimensions that you want to plot. The default value is 10. You can set to other integer values. Note that ndim must be smaller that the no. of brands in your distance matrix.}
}
\value{
A scree plot that can help you decide that no. of dimensions.
}
\description{
The function creates a scree plot for multi-dimensional scaling. A scree plot is a plot that plots the STRESS (standardized residual sum of square) against the no. of dimensions. Using the scree plot, you can apply the elbow criteria to select the no. of dimensions.
}
\examples{
NA
}
