% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sticky_notes}
\alias{sticky_notes}
\title{Sticky Notes}
\format{
A data frame with 170 rows and 2 variables:
\describe{
\item{time}{Each day}
\item{no_of_adoptions}{The number of adoptions (installations)}
}
}
\source{
{Compiled by Xi Chen}
}
\usage{
sticky_notes
}
\description{
Contains the no. of installations on each day.
}
\examples{
data(sticky_notes)
}
\keyword{datasets}
