% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{tajd}
\alias{tajd}
\title{Calculate Tajima's D}
\usage{
tajd(pi, s, n)
}
\arguments{
\item{pi}{theta pi, or pairwise differences}

\item{s}{numbr of segregating sites}

\item{n}{number of samples}
}
\description{
Calculate Tajima's D
}
