% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{theta_pi}
\alias{theta_pi}
\title{Calculate Pi, via the sum of site heterozygosities}
\usage{
theta_pi(x)
}
\arguments{
\item{x}{a gamete matrix}
}
\description{
Calculate Pi, via the sum of site heterozygosities
}

