% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{theta_W}
\alias{theta_W}
\title{Calculate Watterson's Theta}
\usage{
theta_W(s, n)
}
\arguments{
\item{s}{number of segregating sites}

\item{n}{number of samples}
}
\description{
Calculate Watterson's Theta
}
