% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{vif}
\alias{vif}
\title{vif}
\usage{
vif(vnames, data)
}
\arguments{
\item{vnames}{a character vector of variable names of the variables whose VIFs you want to calculate.}

\item{data}{a data frame containing the actual variables.}
}
\value{
This function returns the variance-inflation-factor of a set of independent variables.
}
\description{
The vif function calculates the Variance-Inflation-Factor of a set of independent variables. The rule-of-thumb is that the vif of an IV should be smaller than 10. The function takes two inputs: "vnames" and "data". Note that "vnames" must be consistent with the variable names in "data". Or else, you get an error.
}
\examples{
NA
}
