% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ms.r
\name{write_tee}
\alias{write_tee}
\title{A "tee" (in Unix sense) that writes MS's results to file}
\usage{
write_tee(x, con, pass = TRUE)
}
\arguments{
\item{x}{MS results @param con to save output to}

\item{pass}{if TRUE, pass the input directly to output for pipe operations

This writes a connection \code{con}, then return results (so can be used in
pipe).}
}
\description{
A "tee" (in Unix sense) that writes MS's results to file
}
\examples{
# write MS results to 'test.ms', then pass down the pipeline
\dontrun{
res <- call_ms(10, 500, "-t 5") \%>\% write_tee('test.ms') \%>\% 
             parse_ms() \%>\% mutate(pi=map_dbl(gametes, pi))
}
}
