% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ms.r
\name{ms}
\alias{ms}
\title{Call MS and Parse Output from R}
\usage{
ms(nsam, howmany, cmd = NULL, ..., ms = "ms", verbose = TRUE)
}
\arguments{
\item{nsam}{number of samples (gametes) to draw}

\item{howmany}{how many replicates to run}

\item{cmd}{the command to pass to MS}

\item{...}{command line arguments as function arguments (if not using \code{cmd})}

\item{ms}{the ms-like executable to run, must be in \code{$PATH} or path to executable}
}
\description{
Call Hudon's MS from R, returning a tibble of results that has been parsed
by parse_ms(). This is simply a wrapper for call_ms() and parse_ms()
}
