% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ms.r
\name{parse_keyvals}
\alias{parse_keyvals}
\title{Parse MS's key/value pairs, e.g. segsites and positions
returning a list of key/vals (where vals can be list too)}
\usage{
parse_keyvals(x)
}
\description{
Parse MS's key/value pairs, e.g. segsites and positions
returning a list of key/vals (where vals can be list too)
}
\keyword{internal}
