% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{theta_pi}
\alias{theta_pi}
\title{Calculate Pi, via the sum of site heterozygosities}
\usage{
theta_pi(x, replace = FALSE)
}
\arguments{
\item{x}{a gamete matrix}

\item{replace}{calculate theta pi based on sampling with replacement formula}
}
\description{
Calculate Pi, via the sum of site heterozygosities
}
